fix for the "pick activity empty result" bug at picking a ringtone
I tried to pick a new ringtone over Prospector and always got "pick activity empty result" from the settings app.

With this fix it works now. Main problem was that the settings app requires a Object with a blob and an name property and not just a blob.

I tested it with ringtone and wallpaper picking, both worked quite well.